Code Cleanup

+++ Essentials/src/net/ess3/craftbukkit/SetExpFix.java
@@ -74,7 +74,7 @@ public class SetExpFix
 	//Without this people would be able to use exp and then still sell it.
 	public static int getTotalExperience(final Player player)
 	{
-		int exp = (int)Math.round(getExpAtLevel(player) * player.getExp());
+		int exp = Math.round(getExpAtLevel(player) * player.getExp());
 		int currentLevel = player.getLevel();
 
 		while (currentLevel > 0)
@@ -87,7 +87,7 @@ public class SetExpFix
 
 	public static int getExpUntilNextLevel(final Player player)
 	{
-		int exp = (int)Math.round(getExpAtLevel(player) * player.getExp());
+		int exp = Math.round(getExpAtLevel(player) * player.getExp());
 		int nextLevel = player.getLevel();
 		return getExpAtLevel(nextLevel) - exp;
 	}
